+++ lib/ui/home/HomeRecentItemsWidget.dart
@@ -38,7 +38,7 @@ import 'package:illinois/ui/athletics/AthleticsNewsArticlePanel.dart';
 import 'package:illinois/ui/events/CompositeEventsDetailPanel.dart';
 import 'package:illinois/ui/explore/ExploreDetailPanel.dart';
 import 'package:illinois/ui/guide/GuideDetailPanel.dart';
-import 'package:illinois/ui/widgets/SectionTitlePrimary.dart';
+import 'package:rokwire_plugin/ui/widgets/section_heading.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 
@@ -133,10 +133,10 @@ class _RecentItemsList extends StatelessWidget{
     String? moreLabel = StringUtils.isEmpty(moreButtonLabel)? Localization().getStringEx('widget.home_recent_items.button.more.title', 'View All'): moreButtonLabel;
     return items!=null && items!.isNotEmpty? Column(
       children: <Widget>[
-        SectionTitlePrimary(
+        SectionHeading(
             title:heading,
             subTitle: subTitle,
-            iconPath: headingIconRes,
+            titleIconAsset: headingIconRes,
             children: _buildListItems(context)
         ),
         !showMoreButton?Container():